changing folder location
This update will change the folder location. now all similar components
will be grouped into one place.
