+++ UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/ItemTouchListenerAdapter.java
@@ -47,7 +47,9 @@ public  class ItemTouchListenerAdapter extends GestureDetector.SimpleOnGestureLi
     private RecyclerView recyclerView;
     private GestureDetector gestureDetector;
 
-    public ItemTouchListenerAdapter(RecyclerView recyclerView, RecyclerViewOnItemClickListener listener) {
+    public ItemTouchListenerAdapter(
+            RecyclerView recyclerView,
+            RecyclerViewOnItemClickListener listener) {
         if (recyclerView == null || listener == null) {
             throw new IllegalArgumentException("RecyclerView and Listener arguments can not be null");
         }
@@ -64,6 +66,7 @@ public  class ItemTouchListenerAdapter extends GestureDetector.SimpleOnGestureLi
 
     @Override
     public void onTouchEvent(RecyclerView recyclerView, MotionEvent motionEvent) {
+
     }
 
     @Override
@@ -101,7 +104,8 @@ public  class ItemTouchListenerAdapter extends GestureDetector.SimpleOnGestureLi
         if (view == null) return false;
 
         view.setPressed(false);
-        int position = shiftAdjustInt(recyclerView.getChildPosition(view));
+        int position = shiftAdjustInt(recyclerView.getChildAdapterPosition(view));
+
         if (position != AdmobAdapter.POSITION_ON_AD) {
             listener.onItemClick(recyclerView, view, position);
         }
@@ -111,7 +115,7 @@ public  class ItemTouchListenerAdapter extends GestureDetector.SimpleOnGestureLi
     public void onLongPress(MotionEvent e) {
         View view = getChildViewUnder(e);
         if (view == null) return;
-        int position = shiftAdjustInt(recyclerView.getChildPosition(view));
+        int position = shiftAdjustInt(recyclerView.getChildAdapterPosition(view));
         if (position != AdmobAdapter.POSITION_ON_AD) {
             listener.onItemLongClick(recyclerView, view, position);
         }
