+++ UltimateRecyclerView/ultimaterecyclerview/src/main/java/com/marshalchen/ultimaterecyclerview/layoutmanagers/ScrollSmoothLineaerLayoutManager.java
@@ -1,4 +1,4 @@
-package com.marshalchen.ultimaterecyclerview.uiUtils;
+package com.marshalchen.ultimaterecyclerview.layoutmanagers;
 
 import android.content.Context;
 import android.graphics.PointF;
@@ -19,14 +19,26 @@ public class ScrollSmoothLineaerLayoutManager extends LinearLayoutManager {
     }
 
     /**
-     * {@inheritDoc}
-     *
-     * @param recycler
-     * @param state
+     * @param recycler the object
+     * @param state    the state
      */
     @Override
     public void onLayoutChildren(RecyclerView.Recycler recycler, RecyclerView.State state) {
-        super.onLayoutChildren(recycler, state);
+        try {
+            super.onLayoutChildren(recycler, state);
+        } catch (Exception e) {
+
+        }
+    }
+
+    /**
+     * try to fix the inconsistency detection issue
+     *
+     * @return boolean not support on V23.1.1
+     */
+    @Override
+    public boolean supportsPredictiveItemAnimations() {
+        return false;
     }
 
     @Override
