+++ UltimateRecyclerView/ultimaterecyclerview/src/main/res/layout/bottom_progressbar.xml
@@ -2,7 +2,10 @@
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
+
     <ProgressBar
-        android:layout_width="wrap_content" android:id="@+id/bottom_progress_bar"
-        android:layout_height="wrap_content" android:layout_centerHorizontal="true"/>
+        android:id="@+id/bottom_progress_bar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true" />
 </RelativeLayout>
